# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger для АС ИДК.",
    "version": "0.1.0",
    "description": " ФЯО ФГУП ГХК, ДИТ 2023 г. ",
    "license": {
      "name": "ДИТ",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Ссылка на гит",
      "url": "https://github.com/zhiraff/idc-back"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/v1/personnel/docs": {
      "post": {
        "tags": [
          "personnel"
        ],
        "description": "Создать запись в таблице документы у физ.лица",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "flKey": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "serial": {
                  "example": "any"
                },
                "number": {
                  "example": "any"
                },
                "dateIssue": {
                  "example": "any"
                },
                "whoIssue": {
                  "example": "any"
                },
                "podrIssue": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "personnel"
        ],
        "description": "Показ всех документов",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/v1/personnel/docs/search": {
      "get": {
        "tags": [
          "personnel"
        ],
        "description": "Поиск документа",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "flKey",
            "in": "query",
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "serial",
            "in": "query",
            "type": "string"
          },
          {
            "name": "number",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dateIssue",
            "in": "query",
            "type": "string"
          },
          {
            "name": "whoIssue",
            "in": "query",
            "type": "string"
          },
          {
            "name": "podrIssue",
            "in": "query",
            "type": "string"
          },
          {
            "name": "active",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/personnel/docs/{id}": {
      "get": {
        "tags": [
          "personnel"
        ],
        "description": "Просмотр записи подробно",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "personnel"
        ],
        "description": "Обновить запись в таблице документов физ.лица",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "serial": {
                  "example": "any"
                },
                "number": {
                  "example": "any"
                },
                "dateIssue": {
                  "example": "any"
                },
                "whoIssue": {
                  "example": "any"
                },
                "podrIssue": {
                  "example": "any"
                },
                "active": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "personnel"
        ],
        "description": "Удалить запись в таблице документов у физ.лица",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/v1/document/header/": {
      "get": {
        "tags": [
          "document"
        ],
        "description": "Показ все заголовки документа",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "document"
        ],
        "description": "Создание записи",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "typeDocument": {
                  "example": "any"
                },
                "typeExam": {
                  "example": "any"
                },
                "dateDocument": {
                  "example": "any"
                },
                "numberDocument": {
                  "example": "any"
                },
                "beginPeriod": {
                  "example": "any"
                },
                "endPeriod": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/document/header/search": {
      "get": {
        "tags": [
          "document"
        ],
        "description": "Поиск заголовка документа",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "perpage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "organization",
            "in": "query",
            "type": "string"
          },
          {
            "name": "typeDocument",
            "in": "query",
            "type": "string"
          },
          {
            "name": "typeExam",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dateDocument",
            "in": "query",
            "type": "string"
          },
          {
            "name": "numberDocument",
            "in": "query",
            "type": "string"
          },
          {
            "name": "beginPeriod",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endPeriod",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/document/header/{id}": {
      "get": {
        "tags": [
          "document"
        ],
        "description": "Просмотр заголовка документа подробно",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "document"
        ],
        "description": "Обновление записи",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "organization": {
                  "example": "any"
                },
                "typeDocument": {
                  "example": "any"
                },
                "typeExam": {
                  "example": "any"
                },
                "dateDocument": {
                  "example": "any"
                },
                "numberDocument": {
                  "example": "any"
                },
                "beginPeriod": {
                  "example": "any"
                },
                "endPeriod": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "document"
        ],
        "description": "Удаление записи",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}